apply review requested changes
