+++ BrightID/src/utils/verifications.ts
@@ -14,3 +14,20 @@ export const isVerified = (verifications: Value, verification: string) => {
     return false;
   }
 };
+
+export const getVerificationsTexts = (verifications: Verification[]) => {
+  const texts = [];
+  let v = verifications.find(v => v.name === 'SeedConnected');
+  if (v && v.rank > 0) {
+    texts.push(`Joined Meets`);
+  }
+  v = verifications.find(v => v.name === 'Bitu');
+  if (v && v.score > 0) {
+    texts.push(`Bitu ${v.score}`);
+  }
+  v = verifications.find(v => v.name === 'Seed');
+  if (v) {
+    texts.push('Seed');
+  }
+  return texts;
+}
